Splitted reducers logic
